+++ src/exporters/lootnika_text/formats/json.py
@@ -1,18 +1,19 @@
+"""
+    json serialization for Lootnika document.
+    Create parcel (final external document) for exporter.
+"""
+
 from taskstore import Document
+from lootnika import orjson
 
 
 class Converter:
-    """
-    Convert internal format (lootnika document) into text.
-    Create parcel (final external document) for exporter.
-    """
-
-    def __init__(self, cfg=None):
+    def __init__(self, cfgSection: dict=None, cfgExporter: dict=None):
         """
         Converter must have:
             self.type - self name that can say about output format\n
-            self.encoding - for bytearray(). Use None if you parcel is also like that
-        :param cfg: exporter section, no needed.
+        :param cfgSection: exporter section raw, no needed.
+        :param cfgExporter: exporter validated configuration, no needed.
         """
         self.type = "json"
         self.adds = {"DOCUMENTS": []}
@@ -26,6 +27,6 @@ class Converter:
         will added to new parcel.
         :return: finished parcel. ready to export
         """
-        parcel = f"{self.adds}\n\n"
+        parcel = f"{orjson.dumps(self.adds).decode()}\n\n"
         self.adds = {"DOCUMENTS": []}
         return parcel